requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedgen==1.0.0
python-dateutil==2.8.2
lxml==4.9.3 